tt_symbiote: added custom dispatching
