+++ models/tt_symbiote/core/dispatchers/__init__.py
@@ -1,0 +1,15 @@
+"""Dispatcher implementations for TTNN operation handling."""
+
+from models.tt_symbiote.core.dispatchers.dispatcher_config import (
+    get_active_dispatcher,
+    set_dispatcher,
+    list_available_dispatchers,
+    register_dispatcher,
+)
+
+__all__ = [
+    "get_active_dispatcher",
+    "set_dispatcher",
+    "list_available_dispatchers",
+    "register_dispatcher",
+]
